requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
APScheduler==3.10.4
tiingo==0.14.0
SQLAlchemy==2.0.35
pyodbc==5.2.0
python-dotenv==1.0.1
pandas==2.2.2
requests==2.32.3
dotenv
pydantic_settings>=2.0
pydantic>=2.5
websocket-client
simplejson
